% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnedMean.R
\name{binnedMean}
\alias{binnedMean}
\title{Calculate mean score in each bin.}
\usage{
binnedMean(bins, numvar, mcolname = "score")
}
\arguments{
\item{bins}{Genomic bins GRanges.}

\item{numvar}{Feature scores across genome in RleList.}

\item{mcolname}{Colname of feature score, default = "score".}
}
\value{
Mean score for each bin.
}
\description{
Calculate mean score in each bin.
}
