% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.fit.epigenetic.R
\name{lasso.fit.epigenetic}
\alias{lasso.fit.epigenetic}
\title{1 iteration of LASSO selection on a subset of data.}
\usage{
lasso.fit.epigenetic(x_data, y_data, s)
}
\arguments{
\item{x_data}{Covariates matrix.}

\item{y_data}{Response matrix.}

\item{s}{Seed for random number.}
}
\value{
Epigenetic features selected for 1 iteration of LASSO.
}
\description{
1 iteration of LASSO selection on a subset of data.
}
