% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.to.granges.R
\name{maf.to.granges}
\alias{maf.to.granges}
\title{Convert MAF file to GRanges object.}
\usage{
maf.to.granges(fname.maf)
}
\arguments{
\item{fname.maf}{MAF file.}
}
\value{
GRanges object containing information from MAF file.
}
\description{
Convert MAF file to GRanges object.
}
\examples{
maf.to.granges("SNV_mutations.MAF")
maf.to.granges("Indel_mutations.MAF")
}
