% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCovariate.indel.sparse.R
\name{mutCovariate.indel.sparse}
\alias{mutCovariate.indel.sparse}
\title{Prepare covariate sparse matrix and response matrix.}
\usage{
mutCovariate.indel.sparse(compiled = "mutCovariate-indel-compile-part1.RDS")
}
\arguments{
\item{compiled}{Aggregated table RDS file, default = mutCovariate-compile-part1.RDS.}
}
\value{
A list containing covariates sparse matrix and response matrix.
}
\description{
Prepare covariate sparse matrix and response matrix.
}
