% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.snv.find.motif.R
\name{mutPredict.snv.find.motif}
\alias{mutPredict.snv.find.motif}
\title{Extract nucleotide contexts for each site in SNV hotspots.}
\usage{
mutPredict.snv.find.motif(seq, motifs)
}
\arguments{
\item{seq}{Dataframe containing DNA sequences of each site in the SNV hotspot.}

\item{motifs}{Selected nucleotide contexts to be extracted.}
}
\value{
Nucleotide context feature matrix.
}
\description{
Extract nucleotide contexts for each site in SNV hotspots.
}
