% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide.selection.R
\name{nucleotide.selection}
\alias{nucleotide.selection}
\title{Select nucleotide context through LASSO for SNVs and filter polyATs from indel mutations.}
\usage{
nucleotide.selection(sampled.sites.snv.file, indel.mutations.file,
  cutoff = 0.9, cores = 1)
}
\arguments{
\item{sampled.sites.snv.file}{Sampled SNV mutated and non-mutated sites RDS file.}

\item{indel.mutations.file}{Indel mutations MAF file.}

\item{cutoff}{Frequency cutoff/threshold to determine nucleotide contexts used in prediction model, ranges from 0.5 to 1, default = 0.90.}

\item{cores}{Number of cores, default = 1.}
}
\value{
A list containing frequency of nucleotide contexts selected through lasso, nucleotide contexts that passed the threshold and polyATs filtered indel mutations.
}
\description{
Select nucleotide context through LASSO for SNVs and filter polyATs from indel mutations.
}
