% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_hits.R
\name{plot_top_hits}
\alias{plot_top_hits}
\title{Plot top hotspots.}
\usage{
plot_top_hits(hotspots.file, fdr.cutoff = 0.01, color.muts = "orange",
  mutations.file, mutation.type, top.no = 3)
}
\arguments{
\item{hotspots.file}{Hotspots generated.}

\item{fdr.cutoff}{FDR cutoff, default = 0.01.}

\item{color.muts}{Color points, default = orange.}

\item{mutations.file}{Mutations found in region of interest MAF file.}

\item{mutation.type}{SNV or indel mutations.}

\item{top.no}{Number of top hotspots to plot, default = 3.}
}
\value{
Top hotspots figures.
}
\description{
Plot top hotspots.
}
