% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.regions.from.roi.chr.R
\name{subtract.regions.from.roi.chr}
\alias{subtract.regions.from.roi.chr}
\title{Remove Region A from Region B for 1 chromosome.}
\usage{
subtract.regions.from.roi.chr(rois, sub, chr = "chr10")
}
\arguments{
\item{rois}{Region B, GRanges object.}

\item{sub}{Region A, GRangesList/GRanges object.}

\item{chr}{Chromosome name.}
}
\value{
Filtered Region B for 1 chromosome.
}
\description{
Remove Region A from Region B for 1 chromosome.
}
