% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.regions.from.roi.R
\name{subtract.regions.from.roi}
\alias{subtract.regions.from.roi}
\title{Remove Region A from Region B.}
\usage{
subtract.regions.from.roi(rois, sub, cores = 1)
}
\arguments{
\item{rois}{Region B, GRanges object.}

\item{sub}{Region A, GRangesList/GRanges object.}

\item{cores}{Number of cores, default = 1.}
}
\value{
Filtered Region B.
}
\description{
Remove Region A from Region B.
}
